more API changes
